backend/todo/urls.py
## backend/todo/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import TaskViewSet, CategoryViewSet, ContextEntryViewSet

router = DefaultRouter()
router.register(r'tasks', TaskViewSet)
router.register(r'categories', CategoryViewSet)
router.register(r'context-entries', ContextEntryViewSet)

urlpatterns = [
    path('', include(router.urls)),
]