backend/requirements.txt
## backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
nltk==3.8.1
requests==2.31.0
python-dateutil==2.8.2
django-cors-headers==3.14.0